+++ app/controllers/admin/assets_controller.rb
@@ -49,7 +49,7 @@ class Admin::AssetsController < AdminController
   end
 
   def fixity_report
-    @fixity_report = FixityReportHelper::FixityReportHelper.new()
+    @fixity_report = FixityReport.new()
   end
 
   def display_attach_form
